fix: remove useless async trait, move dao state to contract
