+++ io/Cargo.toml
@@ -12,4 +12,3 @@ gmeta = { git = "https://github.com/gear-tech/gear.git" }
 gstd = { git = "https://github.com/gear-tech/gear.git", branch = "stable" }
 scale-info = { version = "2", default-features = false }
 parity-scale-codec = { version = "3", default-features = false }
-hashbrown = "0.13.1"
